DTOs, inyeccion por constructor y manejo de errores unificado

+++ backend/src/main/java/com/intiwasi/backend/entity/Categoria.java
@@ -22,5 +22,5 @@ public class Categoria {
     private String nomCategoria;
     
     @Column (name = "Estado", nullable = false)
-    private Integer estado = 1;
+    private Byte estado = (byte) 1;
 }
